@@ -1,84 +1,64 @@
 package io.github.cgi.zabbix.api;
 
-import org.codehaus.jackson.map.ObjectMapper;
-import org.codehaus.jackson.map.SerializationConfig;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-
 
 /**
  * Created by cgi on 01.03.2016.
  * Support Class for converting Objects to JSON
  */
 public class Utils {
-    static ObjectMapper mapper = new ObjectMapper();
-    static ObjectMapper mapperIndent = new ObjectMapper();
+	static ObjectMapper mapper = new ObjectMapper();
 
+	public static byte[] getBytesFromInputStream(InputStream inputStream) {
 
-    public static String toJson(Object o) {
-        String s = null;
-        try {
-            s = mapper.writeValueAsString(o);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return s;
+		ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
 
-    }
+		try {
+			byte[] buffer = new byte[1024];
+			int length;
 
-    public static String toJsonWithIndent(Object o) {
-        String s = null;
-        mapperIndent.getSerializationConfig().enable(SerializationConfig.Feature.INDENT_OUTPUT);
-        try {
-            s = mapperIndent.writeValueAsString(o);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return s;
+			while ((length = inputStream.read(buffer)) != -1) {
+				outputStream.write(buffer, 0, length);
+			}
 
-    }
+			outputStream.flush();
 
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
 
-    public static byte[] getBytesFromInputStream(InputStream inputStream) {
+		return outputStream.toByteArray();
+	}
 
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+	public static String inputStreamToString(byte[] bytes) {
 
-        try{
-        byte[] buffer = new byte[1024];
-        int length;
+		ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+		String string = null;
 
-        while ((length = inputStream.read(buffer)) != -1) {
-            outputStream.write(buffer, 0, length);
-        }
+		try {
 
-        outputStream.flush();
+			outputStream.write(bytes);
+			string = outputStream.toString("UTF-8");
 
-        } catch (IOException e){
-            e.printStackTrace();
-        }
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
 
-        return outputStream.toByteArray();
-    }
+		return string;
+	}
 
-
-    public static String inputStreamToString(byte[] bytes)  {
-
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        String string = null;
-
-        try {
-
-            outputStream.write(bytes);
-            string = outputStream.toString("UTF-8");
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        return string;
-
-    }
+	public static String toJson(Object o) {
+		String s = null;
+		try {
+			s = mapper.writeValueAsString(o);
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return s;
+	}
 
 }
