@@ -1,6 +1,6 @@
 package io.github.cgi.zabbix.api;
 
-import org.codehaus.jackson.JsonNode;
+import com.fasterxml.jackson.databind.JsonNode;
 
 public interface ZabbixApi {
 
